code detail polishment
